pass components as props
